Move oss-only test to its own file
Manual commit to match b236e86030fc46e636d11a4d791aeae81bce95ed
